requirements.txt
git+https://github.com/shvdev1/PS-ML-protos.git@main#egg=grpc_protos
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.33.2
setuptools==80.9.0
typing_extensions==4.15.0
